% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Jammer_Stats.R
\name{jammer_stats}
\alias{jammer_stats}
\title{Jammer Stats}
\usage{
jammer_stats(data)
}
\arguments{
\item{data}{frame of jammer stats}
}
\description{
This function takes a data frame of jammer stats and calculates LJ Percentage and Total Points for each jammer
}
\examples{
jammer_stats(data)
}
\keyword{jammer_stats}
\keyword{roller_derby}

