% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/JammerStats.R
\name{JammerStats}
\alias{JammerStats}
\title{Jammer Stats}
\usage{
JammerStats(data)
}
\arguments{
\item{data}{frame of jammer stats}
}
\description{
This function takes a data frame of jammer stats and calculates LJ Percentage and Total Points for each jammer
}
\examples{
JammerStats(data)
}
\keyword{jammer_stats}
\keyword{roller_derby}

